Warning fixes:         * Makefile.in (flow.o): Depend on recog.h.         * cpplib.h (directive_table): Add missing initializiers.         (finclude): Change type of variable `bsize' to size_t.         * cse.c (rtx_cost): Mark parameter `outer_code' with ATTRIBUTE_UNUSED.         * dwarfout.h (dwarfout_label): Wrap prototype in macro RTX_CODE.         * fix-header.c (lookup_std_proto): Cast the result of `strlen' to         `int' when comparing against one.         (cpp_file_line_for_message): Mark parameter `pfile' with         ATTRIBUTE_UNUSED.         (cpp_fatal): Mark parameter `pfile' with ATTRIBUTE_UNUSED.         * flow.c: Include recog.h.         (sbitmap_copy): Cast arguments 1 & 2 of `bcopy' to (PTR).         * function.c (thread_prologue_and_epilogue_insns): Mark parameter         `f' with ATTRIBUTE_UNUSED.         (reposition_prologue_and_epilogue_notes): Likewise.         * genopinit.c (gen_insn): Cast argument of ctype functions to         `unsigned char'.         * haifa-sched.c: Include recog.h.         (blockage_range): Cast result of UNIT_BLOCKED macro to (int) when         comparing against one.         * libgcc2.a (__throw): Revert ATTRIBUTE_UNUSED change for now.         * mips-tfile.c (parse_end): Cast the argument of ctype function to         `unsigned char'.         (parse_ent): Likewise.         (parse_input): Likewise.         * optabs.c (init_libfuncs): Likewise.         * protoize.c (find_rightmost_formals_list): Likewise.         * recog.h (const_double_operand): Fix typo in prototype.         * tlink.c (scan_linker_output): Cast the argument of ctype         function to `unsigned char'.         * toplev.c (check_lang_option): Cast the result of `strlen' to         `int' when comparing against one.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@23155 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/haifa-sched.c
@@ -166,6 +166,7 @@
 #include "insn-attr.h"
 #include "except.h"
 #include "toplev.h"
+#include "recog.h"
 
 extern char *reg_known_equiv_p;
 extern rtx *reg_known_value;
@@ -2858,7 +2859,7 @@ blockage_range (unit, insn)
   unsigned int blockage = INSN_BLOCKAGE (insn);
   unsigned int range;
 
-  if (UNIT_BLOCKED (blockage) != unit + 1)
+  if ((int) UNIT_BLOCKED (blockage) != unit + 1)
     {
       range = function_units[unit].blockage_range_function (insn);
       /* We only cache the blockage range for one unit and then only if
